✒️Feat: 입력 금액에 대한 유효성 검사를 추가한다.
