+++ docs/README.md
@@ -1,0 +1,17 @@
+// - 1
+// 구입 금액 input / keydown
+this.$amountInput = document.querySelector('[data-props="amount-input"]');
+// 확인 button / click
+this.$confirmButton = document.querySelector('[data-props="confirm-button"]');
+// - 2
+// 구매 개수 span
+this.$countSpan = document.querySelector('[data-props="count-span"]');
+// 번호 보기 button / click
+this.$toggleButton = document.querySelector('[data-props="toggle-button"]');
+// - 3
+// 로또 리스트 ul
+this.$lottoList = document.querySelector('.lotto-list');
+// 로또 래퍼 li
+this.$lottoNumbersWrapper = document.querySelector('.lotto-numbers-wrapper');
+// 로또 넘버 span
+this.$lottoNumbers = document.querySelector('.lotto-numbers');
