+++ src/js/components/AppTemplate.js
@@ -8,7 +8,7 @@ const AppTemplate = () => {
           <label class="mb-2 d-inline-block">구입할 금액을 입력해주세요.
           </label>
           <div class="d-flex">
-            <input type="number" class="w-100 mr-2 pl-2" data-props="amount-input" placeholder="구입 금액" />
+            <input type="text" class="w-100 mr-2 pl-2" data-props="amount-input" placeholder="구입 금액" />
             <button type="button" data-props="confirm-button" class="btn btn-cyan">확인</button>
           </div>
         </form>
