+++ src/js/constants.js
@@ -1,0 +1,1 @@
+export const REGEXP_SPECIAL = /[-+`~!@#$%^&*|\\'\\";:\\/?]/gi;
