+++ src/js/helper/valid.js
@@ -1,0 +1,7 @@
+import { REGEXP_SPECIAL } from '../constants.js';
+
+export const isNull = value => value === null || value === undefined;
+
+export const isEquals = (target1, target2) => target1 === target2;
+
+export const isSpecial = value => REGEXP_SPECIAL.test(value);
